Fix server spec from rochong.space to IP address of server.
rochong.space was provided since (presumably) it is the domain of
the OVH server. However, when Capistrano tries to access rochong.
space, a password is requested and ssh is unsuccessful even when
suspected password is entered. Change specification to IP address
such that Capistrano does not ask for a password and can success-
fully connect to server when running 'cap staging check_write_per
missions."